Abortive/Preventive treatments were added
